+++ e2e/test/constants.ts
@@ -1,0 +1,4 @@
+export const ExtensionKey = "Snowflake";
+export const ExtensionID = "snowflake.snowflake-vsc";
+
+export const CheckpointsConfiguration = "snowflake.snowparkCheckpoints";
