+++ e2e/tsconfig.json
@@ -2,10 +2,10 @@
   "compilerOptions": {
     "target": "ES2022",
     "outDir": "out",
-    "lib": ["ES2022"],
+    "lib": ["ES2020"],
     "moduleResolution": "node",
     "sourceMap": true,
-    "rootDir": "./test",
+    "rootDir": "test",
     "strict": true,
     "types": ["node", "@wdio/mocha-framework", "expect-webdriverio", "wdio-vscode-service"]
   }
